debug

+++ examples/xdp/syn_pad.h
@@ -6,7 +6,7 @@
 #ifndef __SYN_PAD_H
 #define __SYN_PAD_H
 
-const unsigned short SYN_PAD_MIN_BYTES = 40;
+const unsigned int SYN_PAD_MIN_BYTES = 40;
 const unsigned char END_OP_VAL = 0;
 const unsigned char NO_OP_VAL = 1;
 
@@ -15,7 +15,7 @@ struct event {
 	unsigned char status;
 	unsigned long long start;
 	unsigned long long end;
-	unsigned short padding;
+	unsigned int padding;
 };
 
 #endif /* __SYN_PAD_H */
